allow mat data with non double type sr

+++ Raw_data_readers/mat_wc_reader.m
@@ -21,7 +21,7 @@ classdef mat_wc_reader < handle
             if ismember('sr',{finfo.name})  %if is possible, load sr from file; else from set_parameters
                 load(raw_filename,'sr'); 
                 obj.sr_infile = true;
-                obj.sr = sr;
+                obj.sr = double(sr);
             else
                 obj.sr_infile = false;
                 if isfield(par,'sr')
